Refatoração o código para a criação de um componente de FIlterItem
